ONEMPERS-137 Add com.lgi.rdk.HdmiCec WPE Framework plugin
This changeset copies the original org.rdk.HdmiCec plugin to a separate directory and adds the following methods:

enableOneTouchView,
triggerAction,
setCecPingInterval,
getConnectedDevices,
setName,
setOneTouchViewPolicy.
Additionally, the following event is added:

onDevicesChanged.
It has been observed that deactivating the plugin without disabling it first (calling the setEnabled method with "false")
caused WPE Framework to crash. As a solution, the plugin Deinitialize() functions for both the original and new plugins
now try to disable the respective plugin.