ARRISAPP-80 change impl to avoid static gtypes
While using gdbus-codegen is convinient, generated
code is registering static types (g_type_register_static_simple)
and this breaks when the plugin is reactivated.

https://libsoup.org/gobject/gobject-Type-Information.html:
"Static types are never loaded or unloaded at run-time as dynamic
types may be. Static types are created with g_type_register_static()
that gets type specific information passed in via a GTypeInfo
structure."

Switching to implementation based on using GDBusProxy directly;
some of the autogenerated code could be reused.

The old implementation is moved to gdbus-codegen-impl folder;
it can be reenabled with GDBUS_USE_CODEGEN_IMPL on compile time.
Could be used e.g. if we switch to out-of-process, where
static gtypes should not be a problem